+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.1.107](https://github.com/duality-labs/duality-web-app/compare/v0.1.106...v0.1.107) (2023-05-01)
+
+
+### Fixes
+
+* Keplr wallet doesn't always connect on load ([#353](https://github.com/duality-labs/duality-web-app/issues/353)) ([58625bc](https://github.com/duality-labs/duality-web-app/commit/58625bca771a6b1336e54fb2a787c75e5c2b17b7))
+
+## [0.1.106](https://github.com/duality-labs/duality-web-app/compare/v0.1.105...v0.1.106) (2023-05-01)
+
+
+### Features
+
+* update deploys to new Netlify subdomain site ([#355](https://github.com/duality-labs/duality-web-app/issues/355)) ([aa6e1cb](https://github.com/duality-labs/duality-web-app/commit/aa6e1cb1780dede7ba6c743b57aa07a3e99e481a))
+
 ## [0.1.105](https://github.com/duality-labs/duality-web-app/compare/v0.1.104...v0.1.105) (2023-04-06)